운영반영 수정 조치

+++ webapp/mvc/controller/individualWorkState/IndiWorkState.controller.js
@@ -237,9 +237,9 @@ sap.ui.define(
       buildDoughChart(aPlanList) {
         const oDetailModel = this.getViewModel();
         const mPlan = {
-          dUsed: parseInt(aPlanList.Cnt01),
-          dPlan: parseInt(aPlanList.Cnt02),
-          dUnPlan: parseInt(aPlanList.Cnt03),
+          dUsed: parseFloat(aPlanList.Cnt01),
+          dPlan: parseFloat(aPlanList.Cnt02),
+          dUnPlan: parseFloat(aPlanList.Cnt03),
           pUsed: parseFloat(aPlanList.Rte01),
           pPlan: parseFloat(aPlanList.Rte02),
           pUnPlan: parseFloat(aPlanList.Rte03),
@@ -287,9 +287,9 @@ sap.ui.define(
         const oChart = FusionCharts(this.sDoughChartId);
         const oDetailModel = this.getViewModel();
         const mPlan = {
-          dUsed: parseInt(aPlanList.Cnt01),
-          dPlan: parseInt(aPlanList.Cnt02),
-          dUnPlan: parseInt(aPlanList.Cnt03),
+          dUsed: parseFloat(aPlanList.Cnt01),
+          dPlan: parseFloat(aPlanList.Cnt02),
+          dUnPlan: parseFloat(aPlanList.Cnt03),
           pUsed: parseFloat(aPlanList.Rte01),
           pPlan: parseFloat(aPlanList.Rte02),
           pUnPlan: parseFloat(aPlanList.Rte03),
